Extract DashboardEntryPoint component from googlesitekit-dashboard.js.
